src/img.c: Fix cast in get16() to be to short not long.
git-svn-id: file:///home/survex-svn/survex/trunk@3512 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/img.c
@@ -91,7 +91,7 @@ static short
 get16(FILE *fh)
 {
    short w = getc(fh);
-   w |= (long)getc(fh) << 8l;
+   w |= (short)getc(fh) << 8l;
    return w;
 }
